Consolidate rate limiter implementation
